# apps/warmhouse_user_api/api/openapi.yaml
openapi: 3.0.3
info:
    title: WarmHouse User API
    description: API для управления пользователями, домами и устройствами в системе умного дома
    version: 1.0.0
    contact:
        name: WarmHouse Team
        email: support@warmhouse.com
servers:
    - url: http://localhost:8080/api/v1
      description: Локальный сервер разработки
    - url: https://api.warmhouse.com/v1
      description: Продакшн сервер
...
paths:
...
    /register:
        post:
            tags:
                - Auth
            summary: Регистрация нового пользователя
            description: Создает нового пользователя в системе
            operationId: registerUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserRegistrationRequest'
                        example:
                            name: "Иван Петров"
                            email: "ivan@example.com"
                            phone: "+7 (999) 123-45-67"
                            password: "securePassword123"
            responses:
                '201':
                    description: Пользователь успешно зарегистрирован
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserRegistrationResponse'
                            example:
                                user_id: "550e8400-e29b-41d4-a716-446655440000"
                '400':
                    description: Некорректные данные запроса
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            examples:
                                validation_error:
                                    summary: Ошибка валидации
                                    value:
                                        error: "validation_error"
                                        message: "Некорректные данные"
                                        details:
                                            - field: "email"
                                              message: "email должен быть валидным адресом"
                                            - field: "password"
                                              message: "пароль должен содержать минимум 8 символов"
                                email_exists:
                                    summary: Email уже существует
                                    value:
                                        error: "email_exists"
                                        message: "Пользователь с таким email уже зарегистрирован"
                '500':
                    description: Внутренняя ошибка сервера
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: "internal_error"
                                message: "Внутренняя ошибка сервера"
    /login:
        post:
            tags:
                - Auth
            summary: Вход в систему
            description: Аутентификация пользователя
            operationId: loginUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserLoginRequest'
                        example:
                            email: "ivan@example.com"
                            password: "securePassword123"
            responses:
                '200':
                    description: Успешная аутентификация
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserLoginResponse'
                            example:
                                user:
                                    id: "550e8400-e29b-41d4-a716-446655440000"
                                    name: "Иван Петров"
                                    email: "ivan@example.com"
                                    phone: "+7 (999) 123-45-67"
                                    created_at: "2024-01-15T10:30:00Z"
                                    updated_at: "2024-01-15T10:30:00Z"
                                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                expires_at: "2024-01-16T10:30:00Z"
                '400':
                    description: Некорректные данные запроса
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: "validation_error"
                                message: "Некорректные данные"
                                details:
                                    - field: "email"
                                      message: "email должен быть валидным адресом"
                '401':
                    description: Неверные учетные данные
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: "invalid_credentials"
                                message: "Неверный email или пароль"
                '500':
                    description: Внутренняя ошибка сервера
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                            example:
                                error: "internal_error"
                                message: "Внутренняя ошибка сервера"
...
components:
...
    schemas:
...
        UserRegistrationRequest:
            type: object
            required:
                - name
                - email
                - password
            properties:
                name:
                    type: string
                    minLength: 1
                    maxLength: 255
                    description: Имя пользователя
                    example: "Иван Петров"
                email:
                    type: string
                    description: Email адрес пользователя
                    example: "ivan@example.com"
                phone:
                    type: string
                    pattern: "^\\+?[1-9]\\d{1,14}$"
                    description: Номер телефона пользователя (опционально)
                    example: "+7 (999) 123-45-67"
                password:
                    type: string
                    minLength: 8
                    maxLength: 128
                    description: Пароль пользователя
                    example: "securePassword123"
            additionalProperties: false
        UserRegistrationResponse:
            type: object
            required:
                - user_id
            properties:
                user_id:
                    type: string
                    format: uuid
                    description: Уникальный идентификатор пользователя
                    example: "550e8400-e29b-41d4-a716-446655440000"
...
        UserLoginRequest:
            type: object
            required:
                - email
                - password
            properties:
                email:
                    type: string
                    format: email
                    description: Email адрес пользователя
                    example: "ivan@example.com"
                password:
                    type: string
                    description: Пароль пользователя
                    example: "securePassword123"
            additionalProperties: false
        UserLoginResponse:
            type: object
            required:
                - user
            properties:
                user:
                    $ref: '#/components/schemas/User'
            additionalProperties: false

        # User schemas
        User:
            type: object
            required:
                - id
                - name
                - email
                - created_at
                - updated_at
            properties:
                id:
                    type: string
                    format: uuid
                    description: Уникальный идентификатор пользователя
                    example: "550e8400-e29b-41d4-a716-446655440000"
                name:
                    type: string
                    description: Имя пользователя
                    example: "Иван Петров"
                email:
                    type: string
                    description: Email адрес пользователя
                    example: "ivan@example.com"
                phone:
                    type: string
                    description: Номер телефона пользователя
                    example: "+7 (999) 123-45-67"
                created_at:
                    type: string
                    format: date-time
                    description: Дата и время создания пользователя
                    example: "2024-01-15T10:30:00Z"
                updated_at:
                    type: string
                    format: date-time
                    description: Дата и время последнего обновления пользователя
                    example: "2024-01-15T10:30:00Z"
            additionalProperties: false
...
        ErrorResponse:
            type: object
            required:
                - error
                - message
            properties:
                error:
                    type: string
                    description: Код ошибки
                    example: "validation_error"
                message:
                    type: string
                    description: Описание ошибки
                    example: "Некорректные данные"
                details:
                    type: array
                    items:
                        type: object
                        properties:
                            field:
                                type: string
                                description: Поле с ошибкой
                            message:
                                type: string
                                description: Сообщение об ошибке
                    description: Детали ошибок валидации
            additionalProperties: false